* Config scripts now check for its existance before copying... should cut   down on number of recompiles. * Added support for aligning text in BeOSCanvas.

+++ config/config-beos.sh
@@ -1,2 +1,6 @@
 #!/bin/sh
-cp ../../../config/config.beos ../../../config/config.h
+if [ ! -e ../../../config/config.h ]
+then 
+	cp ../../../config/config.beos ../../../config/config.h
+fi
+
